refactor: orders filters dir
