+++ src/page/Orders.tsx
@@ -2,7 +2,7 @@ import { Flex } from "@chakra-ui/react"
 import { getAllNoneGoodOrders, getAllOrders } from "api/order/order"
 import { Container } from "component/Container"
 import { SearchFiltersClearBtn } from "component/customTable/SearchFiltersClearBtn"
-import { OrdersFilters } from "component/order/OrdersFilters"
+import { OrdersFilters } from "component/filter/OrdersFilters"
 import { OrdersTable } from "component/order/OrdersTable"
 import { LoadingPage } from "component/page/LoadingPage"
 import { Page } from "component/page/Page"
